research papers tool/requirements.txt
azure-functions
httpx
feedparser
